Split the secure HTTP client utils to one export per file
