+++ packages/twenty-server/src/engine/core-modules/secure-http-client/constants/allow-all-internal-hosts.constant.ts
@@ -1,0 +1,1 @@
+export const ALLOW_ALL_INTERNAL_HOSTS = '*';
