+++ packages/twenty-server/src/engine/core-modules/secure-http-client/utils/__tests__/is-private-ip.util.spec.ts
@@ -1,7 +1,5 @@
-import {
-  isLinkLocalIp,
-  isPrivateIp,
-} from 'src/engine/core-modules/secure-http-client/utils/is-private-ip.util';
+import { isLinkLocalIp } from 'src/engine/core-modules/secure-http-client/utils/is-link-local-ip.util';
+import { isPrivateIp } from 'src/engine/core-modules/secure-http-client/utils/is-private-ip.util';
 
 describe('isPrivateIp', () => {
   describe('loopback addresses', () => {
