+++ packages/twenty-server/src/engine/core-modules/secure-http-client/utils/normalize-allowed-internal-host.util.ts
@@ -1,0 +1,13 @@
+// Operators paste whatever they have at hand (a bare host, host:port or the
+// full issuer URL); connections are matched on the hostname alone.
+export const normalizeAllowedInternalHost = (entry: string): string => {
+  const trimmed = entry.trim();
+
+  try {
+    return new URL(
+      trimmed.includes('://') ? trimmed : `http://${trimmed}`,
+    ).hostname.replace(/^\[|\]$/g, '');
+  } catch {
+    return trimmed.toLowerCase();
+  }
+};
